todo complete

+++ restful-web-services/src/main/java/com/vishwanath/rest/services/restfulwebservices/todo/TodoHardcodedService.java
@@ -9,7 +9,7 @@ import java.util.List;
 @Service
 public class TodoHardcodedService {
     private static List<Todo> todos = new ArrayList<>();
-    private static int idCounter = 0;
+    private static long idCounter = 0;
 
     static {
         todos.add(new Todo(++idCounter, "Vishwa", "learn anguar", new Date(), false));
